requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
audioop-lts==0.2.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.17.0
fsspec==2025.3.0
gradio==5.20.1
gradio_client==1.7.2
groovy==0.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
Jinja2==3.1.6
jiter==0.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.2.3
openai==1.65.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruff==0.9.10
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
